Rename version.py to comfyui_version.py
