+++ .github/workflows/update-version.yml
@@ -29,9 +29,9 @@ jobs:
         run: |
           python -m pip install --upgrade pip
 
-      - name: Update version.py
+      - name: Update comfyui_version.py
         run: |
-          # Read version from pyproject.toml and update version.py
+          # Read version from pyproject.toml and update comfyui_version.py
           python -c '
           import tomllib
 
@@ -40,8 +40,8 @@ jobs:
               config = tomllib.load(f)
               version = config["project"]["version"]
 
-          # Write version to version.py
-          with open("version.py", "w") as f:
+          # Write version to comfyui_version.py
+          with open("comfyui_version.py", "w") as f:
               f.write("# This file is automatically generated by the build process when version is\n")
               f.write("# updated in pyproject.toml.\n")
               f.write(f"__version__ = \"{version}\"\n")
@@ -53,6 +53,6 @@ jobs:
           git config --local user.email "github-actions@github.com"
           git fetch origin ${{ github.head_ref }}
           git checkout -B ${{ github.head_ref }} origin/${{ github.head_ref }}
-          git add version.py
-          git diff --quiet && git diff --staged --quiet || git commit -m "chore: Update version.py to match pyproject.toml"
+          git add comfyui_version.py
+          git diff --quiet && git diff --staged --quiet || git commit -m "chore: Update comfyui_version.py to match pyproject.toml"
           git push origin HEAD:${{ github.head_ref }}
